src/copilot/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.30.5
asyncpg==0.29.0
pydantic==2.7.1
numpy==1.26.4
httpx==0.27.0
pyaml-env==1.2.1

# indexing
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
haystack-ai==2.2.3
aiohttp==3.9.5
pypdf==4.2.0
trafilatura==1.11.0
